+++ cofactory-1.1/SearchSupplymaterialViewController.m
@@ -191,11 +191,11 @@ static NSString *userCellIdentifier = @"userCell";
                 
                 break;
             case 2:
-                cell.infoLabel.text = [NSString stringWithFormat:@"%ld", self.history.price];
+                cell.infoLabel.text = [NSString stringWithFormat:@"%ld", (long)self.history.price];
                 
                 break;
             case 3:
-                cell.infoLabel.text = [NSString stringWithFormat:@"%ld", self.history.width];
+                cell.infoLabel.text = [NSString stringWithFormat:@"%ld", (long)self.history.width];
                 
                 break;
                 
@@ -225,7 +225,7 @@ static NSString *userCellIdentifier = @"userCell";
                 cell.infoLabel.text = self.history.name;
                 break;
             case 2:
-                cell.infoLabel.text = [NSString stringWithFormat:@"%ld", self.history.price];
+                cell.infoLabel.text = [NSString stringWithFormat:@"%ld", (long)self.history.price];
                 break;
             
             case 3:
